pyproject.toml
[tool.poetry]
name = "hexagonal-api"
version = "0.1.0"
description = "API RESTful con FastAPI y SQLite, con autenticación y autorización (JWT), y pruebas unitarias (pytest). Implementada con Arquitectura Hexagonal, Vertical Slice y Screaming Architecture."
authors = ["lesmesl <c.lesmesl@uniandes.edu.co>"]
readme = "README.md"
packages = [
    { include = "app" }
]
package-mode = false


[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlalchemy = "^2.0.31"
pyjwt = "^2.8.0"
alembic = "^1.13.2"
bcrypt = "^4.2.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.3.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
autoflake = "^2.3.1"

[tool.poetry.scripts]
lint_and_format = "scripts.lint_and_format:main"


[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
in-place = true
recursive = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"